End step index brute force at 40 rather than 88
Analysis by @meemo suggests that it's unnecessary to try any step
indexes higher than 40 to find one with good quality; there's no
perceptible improvement to bruteforcing the entire range, and avoiding
this improves performance significantly (by halving the search space).

Also, bruteforcing above 40 has been observed to create spikes in the
first few samples of the audio and require predictor correction.
By reducing the range to 0-40, there is no spiking and the predictor
does not need to be changed.

Suggested-by: meemo <8729155+meemo@users.noreply.github.com>
Signed-off-by: Lauren Kelly <lauren.kelly@msn.com>